algorithm is implemented

+++ src/main/java/com/altice_labs/labseq/config/RedisConfig.java
@@ -7,20 +7,23 @@ import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.serializer.GenericToStringSerializer;
 
+import java.math.BigInteger;
+
 @Slf4j
 @Configuration
 public class RedisConfig {
 
     @Bean
-    public RedisTemplate<Integer, Integer> redisTemplate(RedisConnectionFactory connectionFactory) {
-        RedisTemplate<Integer, Integer> template = new RedisTemplate<>();
+    public RedisTemplate<Integer, BigInteger> redisTemplate(RedisConnectionFactory connectionFactory) {
+
+        RedisTemplate<Integer, BigInteger> template = new RedisTemplate<>();
         template.setConnectionFactory(connectionFactory);
 
         template.setKeySerializer(new GenericToStringSerializer<>(Integer.class));
-        template.setValueSerializer(new GenericToStringSerializer<>(Integer.class));
+        template.setValueSerializer(new GenericToStringSerializer<>(BigInteger.class));
 
         template.setHashKeySerializer(new GenericToStringSerializer<>(Integer.class));
-        template.setHashValueSerializer(new GenericToStringSerializer<>(Integer.class));
+        template.setHashValueSerializer(new GenericToStringSerializer<>(BigInteger.class));
 
         return template;
     }
